improve dry run output so it's a properly escaped shell script
Pass each of the the generated podman run arguments to shlex.quote()
so that the dry run output is a properly-escaped shell script. This is
useful for using podman-compose to generate a shell script that can run
on another host running a containerized distribution like Fedora CoreOS
where a Python runtime is not available on the host. We can run
podman-compose inside a container, however that podman container
currently cannot create other podman containers.