Add some sanitizing functions to the project name
This commit fixes issue #113 (podman-compose doesn't handle correctly
project names with special characters - ie. the one not allowed in pod
names in podman).

Docker compose behavior is to add a suffix and clean the directory name
(so empty string can't appear), where podman-compose is using only the
directory name. After cleaning, if no other characters are found,
the project name is set to a fixed name "invalid_name" to mimick the ability
of docker-compose to work even with exotic directory names.