Add better output to logs and up commands
This commit reimplements the logging part both on "logs" and "up "
commands (to behave better, and also like docker-compose). "up" log
format is now the same code of "logs", which was improved:
* Log lines are prepended with service name
* Log lines are sorted by timestamp

This fixes #111

This code rewrite fixes also the issue #112 where issuing SIGINT was
leaving some containers running. Also the
stacktrace showed when SIGINT is sent to the program during "up"
was fixed because of the rewrite.