Quit when an unknown service operation is reqested
Certain commands like build up down run exec start stop restart logs,
when executed for a service that does not exist, should result in a
human friendly message and exit code 1.

Update the podman_compose.py provide and use service_or_exit function,
that terminates the process if a service with the requested name does
not exist (is not defined).

	% ./podman_compose.py logs xyz
	podman-compose version: 1.0.4
	['podman', '--version', '']
	using podman version: 3.4.4
	Service xyz has not been found.
	% echo $?
	1

Resolve #434

Signed-off-by: Piotr Husiatyński <phusiatynski@gmail.com>