Support building images from a Containerfile
In `normalize_service_final`, podman-compose would arbitrarily set the name of the Containerfile to
"Dockerfile". This made the logic in `build_one` that looks for a number of different names a no-op
(because by the time `build_one` is called, the container dictionary already had an explicit
"Dockerfile").

Podman-compose should support building with a Containerfile (and should probably *prefer* that).

This commit modifies `normalize_service_final` so that it no longer explicitly sets the `dockerfile`
attribute. This allows the existing logic to find the correct file.

Closes #977